[dsx] separated transaction parameters from trade history parameters (because they are subtely different)

+++ xchange-dsx/src/main/java/org/knowm/xchange/dsx/DSXAuthenticatedV2.java
@@ -1,17 +1,5 @@
 package org.knowm.xchange.dsx;
 
-import java.io.IOException;
-import java.math.BigDecimal;
-
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DefaultValue;
-import javax.ws.rs.FormParam;
-import javax.ws.rs.HeaderParam;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
 import org.knowm.xchange.dsx.dto.account.DSXAccountInfoReturn;
 import org.knowm.xchange.dsx.dto.account.DSXCryptoDepositAddressReturn;
 import org.knowm.xchange.dsx.dto.account.DSXCryptoWithdrawReturn;
@@ -21,18 +9,28 @@ import org.knowm.xchange.dsx.dto.account.DSXTransactionStatusReturn;
 import org.knowm.xchange.dsx.dto.trade.DSXActiveOrdersReturn;
 import org.knowm.xchange.dsx.dto.trade.DSXCancelAllOrdersReturn;
 import org.knowm.xchange.dsx.dto.trade.DSXCancelOrderReturn;
+import org.knowm.xchange.dsx.dto.trade.DSXFeesReturn;
 import org.knowm.xchange.dsx.dto.trade.DSXOrder;
 import org.knowm.xchange.dsx.dto.trade.DSXOrderHistoryReturn;
+import org.knowm.xchange.dsx.dto.trade.DSXOrderStatusReturn;
 import org.knowm.xchange.dsx.dto.trade.DSXTradeHistoryReturn;
 import org.knowm.xchange.dsx.dto.trade.DSXTradeReturn;
 import org.knowm.xchange.dsx.dto.trade.DSXTransHistoryResult;
 import org.knowm.xchange.dsx.dto.trade.DSXTransHistoryReturn;
-import org.knowm.xchange.dsx.dto.trade.DSXFeesReturn;
-import org.knowm.xchange.dsx.dto.trade.DSXOrderStatusReturn;
 import org.knowm.xchange.dsx.dto.trade.DSXVolumeReturn;
-
 import si.mazi.rescu.ParamsDigest;
 import si.mazi.rescu.SynchronizedValueFactory;
+
+import javax.ws.rs.Consumes;
+import javax.ws.rs.DefaultValue;
+import javax.ws.rs.FormParam;
+import javax.ws.rs.HeaderParam;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
+import java.io.IOException;
+import java.math.BigDecimal;
 
 /**
  * @author Mikhail Wall
@@ -93,7 +91,7 @@ public interface DSXAuthenticatedV2 extends DSX {
   @POST
   @Path("tapi/v2/history/transactions")
   DSXTransHistoryReturn TransHistory(@HeaderParam("Key") String apiKey, @HeaderParam("Sign") ParamsDigest signer,
-      @FormParam("nonce") SynchronizedValueFactory<Long> nonce, @FormParam("count") Long count,
+      @FormParam("nonce") SynchronizedValueFactory<Long> nonce, @FormParam("count") Integer count,
       @FormParam("fromId") Long fromId, @FormParam("endId") Long endId, @FormParam("order") SortOrder order, @FormParam("since") Long since,
       @FormParam("end") Long end, @FormParam("type") DSXTransHistoryResult.Type type, @FormParam("status") DSXTransHistoryResult.Status status,
       @FormParam("currency") String currency) throws IOException;
@@ -118,7 +116,7 @@ public interface DSXAuthenticatedV2 extends DSX {
   @POST
   @Path("tapi/v2/history/trades")
   DSXTradeHistoryReturn TradeHistory(@HeaderParam("Key") String apiKey, @HeaderParam("Sign") ParamsDigest signer,
-      @FormParam("nonce") SynchronizedValueFactory<Long> nonce, @DefaultValue("1000") @FormParam("count") Long count,
+      @FormParam("nonce") SynchronizedValueFactory<Long> nonce, @DefaultValue("1000") @FormParam("count") Integer count,
       @FormParam("fromId") Long fromId, @FormParam("endId") Long endId, @FormParam("order") SortOrder order, @FormParam("since") Long since,
       @FormParam("end") Long end, @FormParam("pair") String pair) throws IOException;
